requirements.txt
arrow >= 1.2.3
cognite-sdk >= 6.1.2
pandas >= 2.0.1
numpy >= 1.24.3
python-dotenv >= 1.0.0
